# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LEAD CONTROL API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/lead/add": {
            "post": {
                "tags": [
                    "Lead API"
                ],
                "summary": "Add new lead",
                "operationId": "5bcb4c4e077e8a79e93a5e9dc15d4897",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998994630613"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add new lead",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Added successfully!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lead/list": {
            "post": {
                "tags": [
                    "Lead API"
                ],
                "summary": "List lead",
                "operationId": "ddd431610eac4620a6fbe5f6306aeaa6",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List lead",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "new_lead": [],
                                                "my_lead": [
                                                    {
                                                        "id": 1,
                                                        "admin_id": 3,
                                                        "phone": "998994630613",
                                                        "comment": null,
                                                        "status": "my",
                                                        "position": "active",
                                                        "created_at": null,
                                                        "updated_at": "2024-01-12T11:00:33.000000Z"
                                                    },
                                                    {
                                                        "id": 2,
                                                        "admin_id": 3,
                                                        "phone": "998994630614",
                                                        "comment": null,
                                                        "status": "my",
                                                        "position": "active",
                                                        "created_at": null,
                                                        "updated_at": "2024-01-12T11:07:15.000000Z"
                                                    }
                                                ],
                                                "recall_lead": [],
                                                "didntpickup_lead": [],
                                                "coming_lead": []
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lead/detail/{id}": {
            "get": {
                "tags": [
                    "Lead API"
                ],
                "summary": "Get detail lead",
                "operationId": "942f7f23c63afe56f5e7740026226446",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste lead id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check admin **token** and get detail lead.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 2,
                                                "admin_id": 3,
                                                "phone": "998994630614",
                                                "comment": null,
                                                "status": "my",
                                                "position": "active",
                                                "created_at": null,
                                                "updated_at": "2024-01-12T11:07:15.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lead/change": {
            "post": {
                "tags": [
                    "Lead API"
                ],
                "summary": "Change lead",
                "operationId": "4c6c16306d47502970af0a53780b19e4",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lead_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "to_status": {
                                        "type": "string",
                                        "example": "my"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Change lead",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Changed successfully!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}